// backend/src/validation/api.ts
export const GetStatsSummaryQueryParams = zod.object({
  "username": zod.coerce.string().optional()
})
// ...
export const getStatsTrendsQueryDaysDefault = 30;

export const GetStatsTrendsQueryParams = zod.object({
  "username": zod.coerce.string().optional(),
  "mode": zod.coerce.string().optional(),
  "days": zod.coerce.number().default(getStatsTrendsQueryDaysDefault)
})